MAC: set the mac address in netdev_open call from android
netdev_open call is invoked during andorid start and
this will be in user space context. So in this context
sepolicy will allow to read factory partition files.

In general kernel context cannot be given permission to
read the factory partition so user context access is the
right aproach.

move the MAC address specific code to the nv_common.c file

Bug 200470432

Change-Id: I413a92d675eb789d666aafaf0800b20cffc38443
Signed-off-by: Bibhay Ranjan <bibhayr@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1976020
Reviewed-by: Automatic_Commit_Validation_User