Revert "brcmfmac: set mchan to 0"
If mchan is 0 in fw, and when AP connects to DFS
channel, the conencted AP channel flags are
enabled as active but still shows RADAR sensitive
in wl chan_info firmware query
eg below.
Channel 100	A Band, RADAR Sensitive

if channel is marked as RADAR Sensitive, it does
not let AGO creation happen. For AGO creation
to be successfull below is the channel flags on
channel 100 for eg.

Channel 100	A Band

This reverts commit 6b3bfd9a853c71be513bcfd5d2affcdd992325f2.

Bug 200514835

Change-Id: I85c3b93d8877bf2dd5f4dd5ebde1418d53881f21
Reviewed-on: https://git-master.nvidia.com/r/2108426
Tested-by: Bibhay Ranjan <bibhayr@nvidia.com>
Reviewed-by: Mohan Thadikamalla <mohant@nvidia.com>
Reviewed-by: Om Prakash Singh <omp@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bibhay Ranjan <bibhayr@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>